+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/Service.java
@@ -24,7 +24,7 @@ import io.swagger.annotations.ApiModel;
 @ApiModel(value = "Service", description = "This describes one service in the dockstore as a special degenerate case of a workflow")
 @Entity
 @Table(name = "service")
-@NamedQuery(name = "io.dockstore.webservice.core.Service.findAllPublishedPaths", query = "SELECT new io.dockstore.webservice.core.WorkflowPath(c.sourceControl, c.organization, c.repository, c.workflowName) from Service c where c.isPublished = true")
+@NamedQuery(name = "io.dockstore.webservice.core.Service.findAllPublishedPaths", query = "SELECT new io.dockstore.webservice.core.database.WorkflowPath(c.sourceControl, c.organization, c.repository, c.workflowName) from Service c where c.isPublished = true")
 public class Service extends Workflow {
 
     public enum SubClass { DOCKER_COMPOSE, SWARM, KUBERNETES, HELM }
