+++ dockstore-webservice/src/main/java/io/dockstore/webservice/jdbi/BioWorkflowDAO.java
@@ -18,8 +18,8 @@ package io.dockstore.webservice.jdbi;
 import java.util.List;
 
 import io.dockstore.webservice.core.BioWorkflow;
-import io.dockstore.webservice.core.RSSWorkflowPath;
-import io.dockstore.webservice.core.WorkflowPath;
+import io.dockstore.webservice.core.database.RSSWorkflowPath;
+import io.dockstore.webservice.core.database.WorkflowPath;
 import org.hibernate.SessionFactory;
 
 import static io.dockstore.webservice.resources.MetadataResource.RSS_ENTRY_LIMIT;
