+++ dockstore-webservice/src/main/java/io/dockstore/webservice/resources/MetadataResource.java
@@ -55,13 +55,13 @@ import io.dockstore.webservice.DockstoreWebserviceConfiguration;
 import io.dockstore.webservice.api.Config;
 import io.dockstore.webservice.core.Collection;
 import io.dockstore.webservice.core.Organization;
-import io.dockstore.webservice.core.RSSEntryPath;
-import io.dockstore.webservice.core.RSSToolPath;
-import io.dockstore.webservice.core.RSSWorkflowPath;
 import io.dockstore.webservice.core.Tool;
-import io.dockstore.webservice.core.ToolPath;
 import io.dockstore.webservice.core.Workflow;
-import io.dockstore.webservice.core.WorkflowPath;
+import io.dockstore.webservice.core.database.RSSEntryPath;
+import io.dockstore.webservice.core.database.RSSToolPath;
+import io.dockstore.webservice.core.database.RSSWorkflowPath;
+import io.dockstore.webservice.core.database.ToolPath;
+import io.dockstore.webservice.core.database.WorkflowPath;
 import io.dockstore.webservice.helpers.MetadataResourceHelper;
 import io.dockstore.webservice.helpers.PublicStateManager;
 import io.dockstore.webservice.helpers.statelisteners.SitemapListener;
